requirements.txt
yt-dlp
python-telegram-bot
spotipy